+++ backstage.rb
@@ -36,7 +36,7 @@ require 'message_processors'
 require 'jobs'
 require 'logs'
 require 'services'
-
+require 'caches'
 
 Backstage.logger.warn "ENV['REQUIRE_AUTHENTICATION'] is not set, *disabling* authentication" unless ENV['REQUIRE_AUTHENTICATION']